requirements.txt
configparser
pyspark
psutil
numpy
fastapi
uvicorn
python-dotenv
